+++ ai-learn-backend/src/main/java/com/earth/online/player/ailearn/question/infrastructure/SystemQuestionRecord.java
@@ -1,5 +1,6 @@
 package com.earth.online.player.ailearn.question.infrastructure;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -12,7 +13,7 @@ public class SystemQuestionRecord {
     private String question;
     private String questionType;
     private String standardAnswer;
-    private Integer importanceScore;
+    private BigDecimal importanceScore;
     private Integer occurrenceCount;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
@@ -49,10 +50,10 @@ public class SystemQuestionRecord {
     public void setStandardAnswer(String standardAnswer) { this.standardAnswer = standardAnswer; }
 
     /** 获取重要性评分。 */
-    public Integer getImportanceScore() { return importanceScore; }
+    public BigDecimal getImportanceScore() { return importanceScore; }
 
     /** 设置重要性评分。 */
-    public void setImportanceScore(Integer importanceScore) { this.importanceScore = importanceScore; }
+    public void setImportanceScore(BigDecimal importanceScore) { this.importanceScore = importanceScore; }
 
     /** 获取真实面试出现次数。 */
     public Integer getOccurrenceCount() { return occurrenceCount; }
